healthid/api/requirements.txt
fastapi==0.59.0
uvicorn==0.11.5
Jinja2==2.11.2
aiofiles==0.5.0
tinydb==4.1.1
python-multipart==0.0.5
pydantic==1.6.1
python-jose==3.2.0
cryptography==3.0
sse-starlette==0.4.0
fhirclient==3.2.0
